Neuter production bookCleanup for now and add logging (BL-12568)

+++ bookCleanup/bookCleanup.ts
@@ -1,9 +1,12 @@
+import { Context } from "@azure/functions";
 import BloomParseServer from "../common/BloomParseServer";
 import { deleteFilesByPrefix } from "../common/s3";
 import { Environment } from "../common/utils";
 
-export async function bookCleanupInternal(env: Environment) {
+export async function bookCleanupInternal(env: Environment, context: Context) {
   const parseServer = new BloomParseServer(env);
+
+  const runInSafeMode = env === Environment.PRODUCTION;
 
   const sessionToken = await parseServer.loginAsBookCleanupUser();
   const cutoff = Date.now() - 24 * 60 * 60 * 1000; // 1 day ago
@@ -12,16 +15,43 @@ export async function bookCleanupInternal(env: Environment) {
   ).books;
   for (const book of booksToBeCleanedUp) {
     const bookPrefixToDelete = `${book.objectId}/${book.uploadPendingTimestamp}`;
-    await deleteFilesByPrefix(bookPrefixToDelete, env);
+    if (!runInSafeMode) {
+      // Delete files from S3 for partial upload.
+      await deleteFilesByPrefix(bookPrefixToDelete, env);
+    }
+    context.log(
+      `${
+        runInSafeMode ? "Safe Mode. Would have deleted" : "Deleted"
+      } files with prefix ${bookPrefixToDelete} from S3.`
+    );
+
     if (book.baseUrl === undefined) {
-      await parseServer.deleteBookRecord(book.objectId, sessionToken);
+      if (!runInSafeMode) {
+        // Delete new book record which was never fully created.
+        await parseServer.deleteBookRecord(book.objectId, sessionToken);
+      }
+      context.log(
+        `${
+          runInSafeMode ? "Safe Mode. Would have deleted" : "Deleted"
+        } book record with ID ${book.objectId}.`
+      );
     } else {
-      await parseServer.modifyBookRecord(
-        book.objectId,
-        {
-          uploadPendingTimestamp: null,
-        },
-        sessionToken
+      if (!runInSafeMode) {
+        // Update book record to remove uploadPendingTimestamp.
+        await parseServer.modifyBookRecord(
+          book.objectId,
+          {
+            uploadPendingTimestamp: null,
+          },
+          sessionToken
+        );
+      }
+      context.log(
+        `${
+          runInSafeMode ? "Safe Mode. Would have updated" : "Updated"
+        } book record with ID ${
+          book.objectId
+        } to remove uploadPendingTimestamp.`
       );
     }
   }
